Delete comments, move isSubstring function
