+++ src/app/helpers/filter-check.ts
@@ -1,0 +1,3 @@
+export function isSubstring(str: string | number, substr: string): boolean {
+  return str.toString().toLowerCase().includes(substr.toLowerCase());
+}
