+++ src/app/models/coach-profile/coach-profile-table/coach-profile-table.component.html
@@ -1,8 +1,4 @@
 <div class="mat-elevation-z8">
-  <!--  <mat-form-field appearance="standard">-->
-  <!--    <mat-label>Filter</mat-label>-->
-  <!--    <input matInput (keyup)="applyFilter($event)" placeholder="Enter test number" #input />-->
-  <!--  </mat-form-field>-->
   <table mat-table [dataSource]="dataSource" matSort matSortActive="date" matSortStart="asc">
     <ng-container matColumnDef="id">
       <th class="header" mat-header-cell *matHeaderCellDef>
