+++ src/app/models/hr-profile/hr-profile.component.html
@@ -1,6 +1,5 @@
 <div class="mat-elevation-z8">
   <table mat-table [dataSource]="dataSource" matSort>
-    <!-- ID Column -->
     <ng-container matColumnDef="id">
       <th class="header" mat-header-cell *matHeaderCellDef>
         <mat-form-field class="filter" floatLabel="never">
@@ -12,7 +11,6 @@
       <td mat-cell *matCellDef="let row">{{ row.id }}</td>
     </ng-container>
 
-    <!-- Name Column -->
     <ng-container matColumnDef="firstName">
       <th class="header" mat-header-cell *matHeaderCellDef>
         <mat-form-field class="filter" floatLabel="never">
@@ -24,7 +22,6 @@
       <td mat-cell *matCellDef="let row">{{ row.firstName }}</td>
     </ng-container>
 
-    <!-- Last Name Column -->
     <ng-container matColumnDef="lastName">
       <th class="header" mat-header-cell *matHeaderCellDef>
         <mat-form-field class="filter" floatLabel="never">
@@ -36,14 +33,13 @@
       <td mat-cell *matCellDef="let row">{{ row.lastName }}</td>
     </ng-container>
 
-    <!-- Assessment  Column -->
     <ng-container matColumnDef="assessment">
       <th mat-header-cell *matHeaderCellDef mat-sort-header>Assign Test</th>
       <td mat-cell *matCellDef="let row">
         <button mat-raised-button (click)="onClick(row.id)" color="primary">Assign Test</button>
       </td>
     </ng-container>
-    <!-- Info  Column -->
+
     <ng-container matColumnDef="info">
       <th mat-header-cell *matHeaderCellDef mat-sort-header>Information</th>
       <td mat-cell *matCellDef="let row">
